# Conflicts:
#	.env
#	.idea/php.xml
#	composer.lock
#	config/bundles.php
#	src/Controller/TasksController.php
#	src/Controller/TestController.php

+++ src/Controller/FiliereController.php
@@ -9,11 +9,25 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 
-#[Route('admin/filiere')]
+
+#[Route('filiere')]
 class FiliereController extends AbstractController
 {
-    #[Route('/', name: 'app_filiere_index', methods: ['GET'])]
+
+  ////////////////////////////FRONT///////////////////////////////////////
+
+  #[Route('/front', name: 'app_filiere_front', methods: ['GET'])]
+  public function front(FiliereRepository $filiereRepository): Response
+  {
+    
+      return $this->render('filiere/front.html.twig', [
+          'filieres' => $filiereRepository->findAll(),
+      ]);
+  }
+
+    #[Route('/', name: 'app_filiere_index', methods: ['GET']), IsGranted('ROLE_ADMIN')]
     public function index(FiliereRepository $filiereRepository): Response
     {
         $user = $this->getUser();
@@ -27,9 +41,14 @@ class FiliereController extends AbstractController
         ]);
     }
 
-    #[Route('/new', name: 'app_filiere_new', methods: ['GET', 'POST'])]
+    #[Route('/new', name: 'app_filiere_new', methods: ['GET', 'POST']), IsGranted('ROLE_ADMIN')]
     public function new(Request $request, FiliereRepository $filiereRepository): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         $filiere = new Filiere();
         $form = $this->createForm(FiliereType::class, $filiere);
         $form->handleRequest($request);
@@ -46,9 +65,14 @@ class FiliereController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_filiere_show', methods: ['GET'])]
+    #[Route('/{id}', name: 'app_filiere_show', methods: ['GET']), IsGranted('ROLE_ADMIN')]
     public function show(Filiere $filiere): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         $entityManager = $this->getDoctrine()->getManager();
         $filiere = $entityManager->getRepository(Filiere::class)->find($filiere);
         $users = $filiere->getUsers();
@@ -59,9 +83,14 @@ class FiliereController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_filiere_edit', methods: ['GET', 'POST'])]
+    #[Route('/{id}/edit', name: 'app_filiere_edit', methods: ['GET', 'POST']), IsGranted('ROLE_ADMIN')]
     public function edit(Request $request, Filiere $filiere, FiliereRepository $filiereRepository): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         $form = $this->createForm(FiliereType::class, $filiere);
         $form->handleRequest($request);
 
@@ -77,9 +106,14 @@ class FiliereController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_filiere_delete', methods: ['POST'])]
+    #[Route('/{id}', name: 'app_filiere_delete', methods: ['POST']), IsGranted('ROLE_ADMIN')]
     public function delete(Request $request, Filiere $filiere, FiliereRepository $filiereRepository): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         if ($this->isCsrfTokenValid('delete'.$filiere->getId(), $request->request->get('_token'))) {
             $filiereRepository->remove($filiere, true);
         }
@@ -105,5 +139,10 @@ class FiliereController extends AbstractController
         ]);
     }
 
+
+
+
+
+  
 
 }
